endbot: Add RCON unit tests
I also improved the Config file unit tests a little and made it easier
to read the test config file.

Right now, the code inside the event emitters are not tested. I'm not
sure how I can test that code (if it's possible). However, if I figure
it out eventually, it would be nice to do