endbot: Speed up test performance by 75%
I switched from ts-jest to swc-jest. It is much faster because it disables
type checking (the startup time is faster, not the running of the actual
tests. That is still less than 0.1s). Performance went from 4.3s total
to 1.2s on my machine.

I don't think disabling type checking when running tests makes them more
unreliable, but I will revert if that is the case.