docker: Refactor docker containers to use multi-stage builds
This makes the containers a lot easier to maintain

One annoying thing is that 'yarn install' takes 90 seconds to run on my
machine (a problem with yarn, not this configuration). For some reason
it is also the same time for the production and development containers
(but production only installs a few packages).