Update dependencies and require imports to have the '.js' extension
This is a big commit, but it updates all of the dependencies of the
project. Additionally, I now require import statements to have the '.js'
extension because that seems common in many TypeScript projects (e.g.
discord.js). It's annoying that the file extension has to be specified,
but there are many headaches that I had to deal with for trying to avoid
specifying it. It also would make more sense to use the '.ts' extension,
but that requires another compiler to change those to .js, such as
Babel.