Send formatted messages to the Minecraft server
All messages are sent as `[user] message`. The user text will be colored
in Minecraft with the hex color of the Discord user's highest role.

If there are @ pings or # channel declarations in the Discord message,
they will be translated to text before being sent to Minecraft so it
isn't some weird Discord ID in the chat.