Add online command and make it easier to run commands
This is a big commit (I probably should have split it up into multiple,
but oh well). This adds the !online command to retrieve the players
currently online on every server.

I also added a `short_description` to each command which will be shown
in the !help command. The command's description will be shown (in
addition to the short description) when !help <command> is used.

This commit also adds an easier way to run commands. A command context
will be used that contains all necessary information for a Command class
to process the command. I put control of sending the command results in
the command class itself because it allows for more
extensibility--specifically when sending more than one embed, which
allows for waiting for reactions from the user to scroll through more of
the output.